requirements.txt
flask
python-dotenv
smtplib
email-validator
fpdf
python-telegram-bot==13.15
reportlab

